requirements.txt
Django==4.1.0
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2==2.9.3
redis==4.4.0
whitenoise==5.3.0
django-heroku==0.3.1
python-dotenv==0.19.2
python-dateutil
upstash-redis
